+++ rusk/tests/rusk-state.rs
@@ -27,7 +27,7 @@ use ff::Field;
 use parking_lot::RwLockWriteGuard;
 use rand::prelude::*;
 use rand::rngs::StdRng;
-use rusk::node::{Rusk, RuskTip};
+use rusk::node::{Rusk, RuskTip, RuskVmConfig};
 use rusk::Result;
 use tempfile::tempdir;
 use tracing::info;
@@ -39,12 +39,15 @@ const CHAIN_ID: u8 = 0xFA;
 const BLOCK_GAS_LIMIT: u64 = 100_000_000_000;
 const INITIAL_BALANCE: u64 = 10_000_000_000;
 
+const VM_CONFIG: RuskVmConfig =
+    RuskVmConfig::new().with_block_gas_limit(BLOCK_GAS_LIMIT);
+
 // Creates the Rusk initial state for the tests below
 fn initial_state<P: AsRef<Path>>(dir: P) -> Result<Rusk> {
     let snapshot = toml::from_str(include_str!("./config/rusk-state.toml"))
         .expect("Cannot deserialize config");
 
-    new_state(dir, &snapshot, BLOCK_GAS_LIMIT)
+    new_state(dir, &snapshot, VM_CONFIG)
 }
 
 fn leaves_from_height(rusk: &Rusk, height: u64) -> Result<Vec<NoteLeaf>> {
@@ -185,7 +188,8 @@ async fn generate_phoenix_txs() -> Result<(), Box<dyn std::error::Error>> {
     let snapshot = toml::from_str(include_str!("./config/bench.toml"))
         .expect("Cannot deserialize config");
 
-    let rusk = new_state(&tmp, &snapshot, 100_000_000_000)?;
+    let vm_config = RuskVmConfig::new().with_block_gas_limit(100_000_000_000);
+    let rusk = new_state(&tmp, &snapshot, vm_config)?;
 
     let cache =
         Arc::new(std::sync::RwLock::new(std::collections::HashMap::new()));
@@ -247,7 +251,8 @@ async fn generate_moonlight_txs() -> Result<(), Box<dyn std::error::Error>> {
     let snapshot = toml::from_str(include_str!("./config/bench.toml"))
         .expect("Cannot deserialize config");
 
-    let rusk = new_state(&tmp, &snapshot, 100_000_000_000)?;
+    let vm_config = RuskVmConfig::new().with_block_gas_limit(100_000_000_000);
+    let rusk = new_state(&tmp, &snapshot, vm_config)?;
 
     let cache =
         Arc::new(std::sync::RwLock::new(std::collections::HashMap::new()));