+++ rusk/tests/services/contract_stake.rs
@@ -17,6 +17,7 @@ use dusk_vm::gen_contract_id;
 use node_data::ledger::SpentTransaction;
 use rand::prelude::*;
 use rand::rngs::StdRng;
+use rusk::node::RuskVmConfig;
 use rusk::{Result, Rusk};
 use std::collections::HashMap;
 use tempfile::tempdir;
@@ -30,6 +31,9 @@ use crate::common::*;
 
 const BLOCK_HEIGHT: u64 = 1;
 const BLOCK_GAS_LIMIT: u64 = 100_000_000_000;
+const VM_CONFIG: RuskVmConfig =
+    RuskVmConfig::new().with_block_gas_limit(BLOCK_GAS_LIMIT);
+
 const GAS_LIMIT: u64 = 10_000_000_000;
 const GAS_PRICE: u64 = 1;
 
@@ -39,7 +43,7 @@ fn stake_state<P: AsRef<Path>>(dir: P) -> Result<Rusk> {
         toml::from_str(include_str!("../config/stake_from_contract.toml"))
             .expect("Cannot deserialize config");
 
-    new_state(dir, &snapshot, u64::MAX)
+    new_state(dir, &snapshot, VM_CONFIG)
 }
 
 #[tokio::test(flavor = "multi_thread")]